v2.9.79: fix premarket-sim offline mode + empty strategy in matrix + system_force CN
Backend fixes:
- scanner_debug.py: debug/premarket-sim now works WITHOUT scanner running
  Reads from MongoDB directly (stock_daily_ak_full, stock_basic, sentiment_scores, broker_orders)
  Provides full market_snapshot, candidates, strategy_groups, sentiment, hit_rate, limit_pools
  Stock names resolved from stock_basic collection (not stock_daily_ak_full which lacks name field)
  Sentiment period mapping handles both English (RISING/BEARISH) and Chinese (高潮/冰点) values
- scanner_sentiment.py: Empty strategy in broker_orders mapped to 'system_force'
  (was showing as blank '' in strategy×sentiment matrix)

Frontend fixes:
- scanner.ts: Added system_force strategy meta (系统强平) for proper CN display

All 23+ API endpoints verified working. Frontend builds clean (0 monitor TS errors).
Strategy backtest module verified unaffected (separate router, 17 routes intact).

+++ AgentServer/nodes/web/api/scanner_core.py
@@ -550,7 +550,7 @@ async def get_kline_data(ts_code: str, days: int = 30):
         if not mongo_manager.is_initialized:
             return {"success": True, "data": {"ts_code": ts_code, "kline": [], "annotations": []}}
         from datetime import datetime, timedelta
-        start_date = (datetime.now() - timedelta(days=days * 2)).strftime("%Y%m%d")
+        start_date = int((datetime.now() - timedelta(days=days * 2)).strftime("%Y%m%d"))
 
         cursor = mongo_manager.db["stock_daily_ak_full"].find(
             {"ts_code": ts_code, "trade_date": {"$gte": start_date}},
